Move basic project routes to projects blueprint
- Extract projects list and create_project routes from app.py
- Move to projects blueprint with proper /projects URLs
- Import required models and utilities for project management
- Reduced app.py routes from 97 to 95

Progress: Successfully moved template management and basic project routes
to their respective blueprints as part of systematic refactoring.
